integrate adminWidget into group Management
